changed all the numpy types because of deprecation

+++ koogu/data/raw.py
@@ -56,7 +56,7 @@ class Settings:
             assert 0 <= self.win_overlap_samples < self.win_len_samples
 
             if nfft_equals_win_len in [None, False]:
-                self.nfft = int(2 ** np.ceil(np.log2(self.win_len_samples)).astype(np.int))  # nextpow2
+                self.nfft = int(2 ** np.ceil(np.log2(self.win_len_samples)).astype(int))  # nextpow2
             else:
                 self.nfft = self.win_len_samples
 
@@ -299,11 +299,11 @@ class Audio:
                 np.lib.stride_tricks.as_strided(data, shape=shape, strides=strides, writeable=False),
                 np.expand_dims(data[(-clip_len):], 0)], axis=0)
             clip_start_samples = np.concatenate([
-                np.arange(0, len(data) - clip_len + 1, clip_advance, dtype=np.int),
+                np.arange(0, len(data) - clip_len + 1, clip_advance, dtype=int),
                 [(len(data) - clip_len + 1)]], axis=0)
         else:
             sliced_data = np.lib.stride_tricks.as_strided(data, shape=shape, strides=strides, writeable=False)
-            clip_start_samples = np.arange(0, len(data) - clip_len + 1, clip_advance, dtype=np.int)
+            clip_start_samples = np.arange(0, len(data) - clip_len + 1, clip_advance, dtype=int)
 
         return sliced_data, clip_start_samples
 
@@ -377,7 +377,7 @@ class Audio:
 
     @staticmethod
     def __audioread_samp_gen(fh, s_start, s_end, channels, filepath):
-        # audioread produces int16 samples
+        # audioread produces int samples
         n_bytes = 2
         fmt_str = '<i2'
 
@@ -424,7 +424,7 @@ class Audio:
             empty_retval = False    # at least one piece was returned
 
         if empty_retval:
-            yield np.zeros((out_n_channels, 0), dtype=np.int16)
+            yield np.zeros((out_n_channels, 0), dtype=int)
 
     @staticmethod
     def __validate_channels(num_available, requested_idxs):
@@ -453,10 +453,10 @@ class Audio:
 class Convert:
 
     @staticmethod
-    def float2pcm(data, dtype=np.int16):
+    def float2pcm(data, dtype=int):
         """Convert waveform from float to integer. Values outside of [-1.0, 1.0) will get clipped."""
 
-        assert dtype in [np.int16, np.int32]
+        assert dtype in [int, int]
 
         if data.dtype == dtype:
             return data
@@ -600,7 +600,7 @@ class Filters:
         """A 1-dimensional Laplacian of Gaussian kernel for convolutions."""
 
         kernel = Filters.gauss_kernel_1d(sigma)
-        kernel_len = np.int32(len(kernel))
+        kernel_len = int(len(kernel))
         kernel_half_len = kernel_len // 2
 
         # Compute the Laplacian of the above Gaussian.
@@ -610,7 +610,7 @@ class Filters:
                  ((np.arange(-kernel_half_len, kernel_half_len + 1) ** 2) - (sigma ** 2))
 
         # Sum of the points within one-sigma of mean
-        scale_threshold_factor = np.sum(kernel) - (2 * np.sum(kernel[0:np.ceil(2 * sigma).astype(np.int)]))
+        scale_threshold_factor = np.sum(kernel) - (2 * np.sum(kernel[0:np.ceil(2 * sigma).astype(int)]))
         # Note: Doing this before removal of DC (below) because it undesirably lowers the threshold for larger sigma.
 
         # Normalize, in order to set the convolution outputs to be closer to putative blobs' original SNRs
@@ -639,7 +639,7 @@ class Filters:
         melpoints = np.linspace(freq_extents_mel[0], freq_extents_mel[1], num_banks + 2)
 
         # convert mels to Hz and then on to fft bin idx
-        bin_idxs = np.floor((nfft + 1) * (700 * (10 ** (melpoints / 2595.0) - 1)) / fs).astype(np.int)
+        bin_idxs = np.floor((nfft + 1) * (700 * (10 ** (melpoints / 2595.0) - 1)) / fs).astype(int)
 
         filterbank = np.zeros([nfft // 2 + 1, num_banks], dtype=dtype)
         for bank_idx in range(0, num_banks):
